Only output console messages when doing user-initiated download-dev-keys
