+++ bin/next-build-tools.js
@@ -246,5 +246,5 @@ program
 program.parse(process.argv);
 
 if (!process.argv.slice(2).length) {
-program.outputHelp();
+	program.outputHelp();
 }
