+++ lib/development-keys-path.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = process.env.HOME + "/.next-development-keys.json";
