+++ lib/keys.js
@@ -2,9 +2,10 @@
 
 var denodeify = require('denodeify');
 var readFile = denodeify(require('fs').readFile);
+var developmentKeysPath = require('./development-keys-path');
 
 module.exports = function() {
-	return readFile(process.env.HOME + '/.next-development-keys.json', { encoding: 'utf8' })
+	return readFile(developmentKeysPath, { encoding: 'utf8' })
 		.then(function(env) {
 			env = JSON.parse(env);
 			return env;
